// src/api/stetsom/model/postApiCategoriesBody.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiCategoriesBodyName } from "./postApiCategoriesBodyName";
import type { PostApiCategoriesBodySlug } from "./postApiCategoriesBodySlug";

export type PostApiCategoriesBody = {
  name: PostApiCategoriesBodyName;
  slug: PostApiCategoriesBodySlug;
  icon_library_id?: string;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  order?: number;
};